Move similarity matrix computation to `mmr.py`
